backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
requests==2.31.0
pytest-django==4.8.0
